app/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
class-registry==2.1.2
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-model-utils==4.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
filters==1.3.2
gunicorn==20.1.0
idna==2.10
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.3.17
requests==2.25.1
simplejwt==2.0.1
six==1.15.0
sqlparse==0.4.1
typing==3.7.4.3
urllib3==1.26.4
